js: Refactor internals and interfaces
* Collapse `internals/util` into `internals`
* Move interfaces into the toplevel `src` directory
