+++ js/src/index.ts
@@ -1,2 +1,4 @@
 export { AEAD } from "./aead";
 export { SIV } from "./siv";
+
+export * from "./interfaces"
