+++ js/src/internals/block.ts
@@ -1,7 +1,7 @@
 /** Type which represents AES blocks */
 
-import { select } from "./util/constant-time";
-import { wipe } from "./util/wipe";
+import { select } from "./constant-time";
+import { wipe } from "./wipe";
 
 /** An AES block (128-bits) */
 export default class Block {
