+++ js/src/internals/constant-time.ts
