+++ js/src/internals/wipe.ts
