+++ js/src/internals/xor.ts
