+++ js/src/providers/polyfill/aes.ts
@@ -12,8 +12,8 @@
 // (rijndael-alg-fst.c, 3.0, December 2000)
 
 import Block from "../../internals/block";
-import { IBlockCipher } from "../../internals/interfaces";
-import { wipe } from "../../internals/util/wipe";
+import { IBlockCipher } from "../../interfaces";
+import { wipe } from "../../internals/wipe";
 
 // Powers of x mod poly in GF(2).
 const POWX = new Uint8Array([
