+++ js/src/providers/webcrypto.ts
@@ -1,5 +1,5 @@
 import NotImplementedError from "../exceptions/not_implemented_error";
-import { IBlockCipher, ICryptoProvider, ICTRLike } from "../internals/interfaces";
+import { IBlockCipher, ICryptoProvider, ICTRLike } from "../interfaces";
 import WebCryptoAes from "./webcrypto/aes";
 import WebCryptoAesCtr from "./webcrypto/aes_ctr";
